+++ coreloop/lusee_appIds.h
@@ -14,8 +14,8 @@
 // uC generater housekeeping; Priority:  High     
 #define AppID_uC_Housekeeping 0x0206
 
-// Calibrator detected with meta data; Priority:  High     
-#define AppID_Calibrator_Detect 0x0207
+// End of sequence; Priority:  High     
+#define AppID_End_Of_Sequence 0x0207
 
 // Data from the bootloader; Priority:  High     
 #define AppID_uC_Bootloader 0x0208
@@ -25,9 +25,6 @@
 
 // Regular hearbeat packet; Priority:  NONE     
 #define AppID_uC_Heartbeat 0x020A
-
-// For sequencer with a limited number of steps; Priority:  High     
-#define AppID_uC_Sequencer_complete 0x020B
 
 // x = 0..F for 16 correlations; Priority:  High     
 #define AppID_MetaData 0x020F